fixed bugs - duplication of pound sign

+++ database/migrations/2024_10_30_192020_create_boxes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('title');
             $table->longText('description');
             $table->string('type');
-            $table->float('price');
+            $table->decimal('price',6,2);
             $table->string('imagePath');
             $table->timestamps();
